requirements-test.txt
# Test dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Include all production dependencies
-r requirements.txt 